Refactor featured image component
