+++ web/src/app/content/single/single.component.scss
@@ -33,6 +33,10 @@
   }
 }
 
+.post-featured-image {
+  margin: 3.2rem 0;
+}
+
 .horizontal-separator {
   border-top: none;
   border-left: none;
