+++ web/src/app/shared/post/featured-image/featured-image.component.html
@@ -1,0 +1,3 @@
+<a *ngIf="src" [routerLink]="isDisabledLink() ? null : href" [attr.aria-label]="post.title">
+  <img class="featured-image lazyload" [attr.alt]="post.title" [attr.src]="src" />
+</a>
