+++ web/src/app/shared/post/featured-image/featured-image.component.scss
@@ -1,0 +1,8 @@
+:host {
+  display: block;
+}
+
+.featured-image {
+  border-radius: 0.4rem;
+  max-width: 100%;
+}
