+++ web/src/app/shared/post/featured-image/index.ts
@@ -1,0 +1,1 @@
+export * from './featured-image.component';
