+++ web/src/app/shared/post/post.module.ts
@@ -7,10 +7,10 @@ import { StoreModule } from '@ngrx/store';
 import { TemplateModule } from '../template/template.module';
 import { PostAuthorComponent } from './author';
 import { PostCategoriesComponent } from './categories';
+import { PostFeaturedImageComponent } from './featured-image';
 import { PostMetadataComponent } from './metadata';
 import { PostContentComponent } from './post-content.component';
 import { PostExcerptComponent } from './post-excerpt.component';
-import { PostFeaturedImageComponent } from './post-featured-image.component';
 import { PostTagsComponent } from './post-tags.component';
 import { PostTitleComponent } from './post-title.component';
 import * as postReducer from './post.reducer';
